Cleanup

+++ ByteCode.cpp
@@ -667,7 +667,7 @@ void ByteCode::printStack(std::vector<Value> vec)
 	std::cout << '[';
 	if (!vec.empty( ))
 	{
-		for (long long i = 0; i < vec.size( ); i++)
+		for (unsigned int i = 0; i < vec.size( ); i++)
 		{
 			std::cout << vec[i];
 			if (i + 1 != vec.size( ))
